+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - Support Humidifier light. (#184)
 - Expose energy usage for outlets/switches. (#190) Thanks @lstrojny for the contribution
 - Strict config validate for `deviceOverrides`. (#278)
+- Support AirPurifier air quality.
 
 
 ## [1.6.0] - (2022.12.3)